Bugfix/logger (#81)
* Refactored tensorboard.py
Changed log_dir to be mendatory specified
Updated the way log_dir and run_name are joined

* Added WandbCallback

* Renamed WandbCallback to WandbLogger

* Fixed tensorboard test

* Removed wandb.init from callback

* Fixed tensorboard test

* Added wandb test for WandbLogger

* Fixed TensorboardCallback docstring

* Fixed test names

* Added LoggingCallback._get_value
Refactored TensorboardCallback
Refactored WandbLogger

* Fixed extract _from None lgos

* Added wandb to gitignore

* Refactored callbacks.__init__

* Precised imports for tensorboard and wandb

* Restructured tests

* Restructured inits

* Fixed logger with title_on_top for single_agent
Added dummy test for logging visualisation
